# pkg/handlers/auth.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/moxicom/todo-back/models"
)

//	@Summary		SignUp
//	@Tags			auth
//	@Description	create account
//	@ID				create-account
//	@Accept			json
//	@Produce		json
//	@Param			input	body		models.User	true	"account info"
//	@Success		200		{integer}	integer		1
//	@Failure		400,404	{object}	errorMsg
//	@Failure		409		{object}	errorMsg
//	@Failure		default	{object}	errorMsg
//	@Router			/auth/sign-up [post]
func (h *Handler) SignUp(c *gin.Context) {
	var input models.User

	if err := c.BindJSON(&input); err != nil {
		newResponseError(c, http.StatusBadRequest, err.Error())
		return
	}

	// h.Service.Auth.CreateUser...
	id, err := h.service.Auth.CreateUser(input)
	if err != nil {
		newResponseError(c, http.StatusConflict, err.Error())
		return
	}

	c.JSON(http.StatusOK, map[string]interface{}{
		"id": id,
	})
}

type signInInput struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

//	@Summary		SignIn
//	@Tags			auth
//	@Description	sign in to account
//	@ID				sign-in
//	@Accept			json
//	@Produce		json
//	@Param			input	body		signInInput	true	"credentials"
//	@Success		200		{string}	string		"token"
//	@Failure		400,404	{object}	errorMsg
//	@Failure		401		{object}	errorMsg
//	@Failure		default	{object}	errorMsg
//	@Router			/auth/sign-in [post]
func (h *Handler) SignIn(c *gin.Context) {
	var input signInInput

	if err := c.BindJSON(&input); err != nil {
		newResponseError(c, http.StatusInternalServerError, err.Error())
		return
	}

	token, err := h.service.Auth.CreateToken(input.Username, input.Password)
	if err != nil {
		newResponseError(c, http.StatusUnauthorized, err.Error())
		return
	}

	c.JSON(http.StatusOK, map[string]interface{}{
		"token": token,
	})
}